requirements.txt
allure-behave==2.13.1
allure-python-commons==2.13.1
allure-robotframework==2.13.1
annotated-types==0.5.0
anyio==3.7.1
Appium-Python-Client==2.8.1
appnope==0.1.3
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
backoff==2.2.1
behave==1.2.6
blinker==1.7.0
certifi==2022.12.7
chardet==5.2.0
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
cryptography @ file:///private/tmp/python-cryptography-20231024-6375-1k9sys9/cryptography-41.0.5
databricks-cli==0.17.7
decorator==5.1.1
distro==1.9.0
docker==6.1.3
docutils==0.19
eli5==0.13.0
entrypoints==0.4
exa-py==1.0.8
exceptiongroup==1.1.1
executing==1.2.0
Flask==3.0.2
giskard==2.0.0b1
gitdb==4.0.10
GitPython==3.1.32
googleapis-common-protos==1.60.0
graphql-core==3.2.3
graphviz==0.20.1
grpcio==1.51.1
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
hypothesis==6.98.4
hypothesis-graphql==0.11.0
hypothesis-jsonschema==0.23.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonpath-ng==1.5.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
junit-xml==1.9
kitchen==1.2.6
langdetect==1.0.9
llvmlite==0.40.1
lockfile==0.12.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mixpanel==4.10.0
mlflow-skinny==2.5.0
multidict==6.0.5
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
openai==1.11.1
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parse==1.19.0
parse-type==0.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.4.0
ply==3.11
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycryptodome==3.18.0
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
PyJWT==2.8.0
pyngrok==6.0.0
pyrate-limiter==2.10.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==8.0.0
pytest-subtests==0.7.0
python-daemon==2.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.33.0
requests==2.28.2
requests-toolbelt==1.0.0
robotframework==6.0.2
robotframework-appiumlibrary==2.0.0
robotframework-jsonlibrary==0.5
robotframework-pythonlibcore==4.1.2
robotframework-requests==0.9.4
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==6.0.0
rpds-py==0.17.1
schemathesis==3.25.1
scikit-learn==1.3.0
scipy==1.11.1
selenium==4.8.3
shap==0.42.1
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.37.1
starlette-testclient==0.3.0
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.2.0
tomli==2.0.1
tomli_w==1.0.0
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.7.1
urllib3==1.26.15
wcwidth==0.2.6
webdriver-manager==3.8.5
websocket-client==1.6.1
Werkzeug==3.0.1
wsproto==1.2.0
xxhash==3.3.0
yarl==1.9.4
zipp==3.16.2
zstandard==0.21.0
